+++ src/module/checkbox/checkbox.tpl.html
@@ -1,7 +1,7 @@
 <label class="gm-checkbox-wrapper">
-    <span class="gm-radio-checkbox gm-checkbox${parseData.checked === 'checked' ? ' gm-checkbox-checked' : ''}${parseData.checked === 'indeterminate' ? ' gm-checkbox-indeterminate' : ''}">
-        <input type="checkbox" class="gm-radio-checkbox-input gm-checkbox-input"${parseData.value ? ' value="' + parseData.value + '"' : ''}${parseData.checked === 'checked' ? ' checked="true"' : ''}/>
+    <span class="gm-radio-checkbox gm-checkbox{{parseData.checked === 'checked' ? ' gm-checkbox-checked' : ''}}{{parseData.checked === 'indeterminate' ? ' gm-checkbox-indeterminate' : ''}}">
+        <input type="checkbox" class="gm-radio-checkbox-input gm-checkbox-input"{{parseData.value ? ' value="' + parseData.value + '"' : ''}}{{parseData.checked === 'checked' ? ' checked="true"' : ''}}/>
         <span class="gm-checkbox-inner"></span>
     </span>
-    ${parseData.label ? '<span>' + parseData.label + '</span>' : ''}
+    {{parseData.label ? '<span>' + parseData.label + '</span>' : ''}}
 </label>
