+++ src/module/checkbox/column.tpl.html
@@ -1,1 +1,1 @@
-<td gm-checkbox="true" gm-create="true">${parseData.template}</td>
+<td gm-checkbox="true" gm-create="true">{{parseData.template}}</td>
