+++ src/module/config/config.tpl.html
@@ -2,6 +2,6 @@
     <span class="config-action">
         <i class="iconfont icon-close"></i>
     </span>
-    <div class="config-info">${parseData.configInfo}</div>
+    <div class="config-info">{{parseData.configInfo}}</div>
     <ul class="config-list"></ul>
 </div>
